# packages/contracts/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Remnawave AWG internal contracts",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/internal/subscriptions/material": {
      "post": {
        "operationId": "post_SubscriptionMaterial",
        "security": [
          {
            "serviceToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionMaterial"
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionRequest"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Endpoint": {
        "additionalProperties": false,
        "properties": {
          "host": {
            "maxLength": 253,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9.:_-]+$",
            "title": "Host",
            "type": "string"
          },
          "port": {
            "maximum": 65535,
            "minimum": 1,
            "title": "Port",
            "type": "integer"
          }
        },
        "required": [
          "host",
          "port"
        ],
        "title": "Endpoint",
        "type": "object"
      },
...
      "ProtocolSpec": {
        "additionalProperties": false,
        "properties": {
          "adapter_id": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9._-]*$",
            "title": "Adapter Id",
            "type": "string"
          },
          "version": {
            "maxLength": 32,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Version",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ]
            },
            "title": "Parameters",
            "type": "object"
          },
          "required_features": {
            "items": {
              "type": "string"
            },
            "title": "Required Features",
            "type": "array"
          }
        },
        "required": [
          "adapter_id",
          "version"
        ],
        "title": "ProtocolSpec",
        "type": "object"
      },
...
      "SubscriptionPeer": {
        "additionalProperties": false,
        "properties": {
          "profile_id": {
            "format": "uuid",
            "title": "Profile Id",
            "type": "string"
          },
          "node_id": {
            "format": "uuid",
            "title": "Node Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "endpoint": {
            "$ref": "#/components/schemas/Endpoint"
          },
          "server_public_key": {
            "title": "Server Public Key",
            "type": "string"
          },
          "client_private_key": {
            "format": "password",
            "title": "Client Private Key",
            "type": "string",
            "writeOnly": true
          },
          "ipv4_address": {
            "title": "Ipv4 Address",
            "type": "string"
          },
          "ipv6_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Ipv6 Address"
          },
          "dns_servers": {
            "items": {
              "type": "string"
            },
            "title": "Dns Servers",
            "type": "array"
          },
          "allowed_ips": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Ips",
            "type": "array"
          },
          "mtu": {
            "default": 1420,
            "title": "Mtu",
            "type": "integer"
          },
          "protocol": {
            "$ref": "#/components/schemas/ProtocolSpec"
          }
        },
        "required": [
          "profile_id",
          "node_id",
          "name",
          "endpoint",
          "server_public_key",
          "client_private_key",
          "ipv4_address",
          "protocol"
        ],
        "title": "SubscriptionPeer",
        "type": "object"
      },
      "SubscriptionMaterial": {
        "additionalProperties": false,
        "properties": {
          "user_uuid": {
            "format": "uuid",
            "title": "User Uuid",
            "type": "string"
          },
          "peers": {
            "items": {
              "$ref": "#/components/schemas/SubscriptionPeer"
            },
            "title": "Peers",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          }
        },
        "required": [
          "user_uuid",
          "peers"
        ],
        "title": "SubscriptionMaterial",
        "type": "object"
      },
      "SubscriptionRequest": {
        "additionalProperties": false,
        "properties": {
          "subscription_token": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "title": "Subscription Token",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "subscription_token"
        ],
        "title": "SubscriptionRequest",
        "type": "object"
      },
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "APIError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "mutualTLS": {
        "type": "mutualTLS"
      },
      "serviceToken": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}